requirements.txt
Django==4.1.4
psycopg2==2.9.5
requests==2.28.2
django-bootstrap-v5==1.0.11
django-import-export==3.0.2